+++ CHANGELOG.md
@@ -9,6 +9,8 @@ next
 * Default values of `false` now work correctly.
 * Added `BigDecimal`, `Date` and `Time` types
 * `DateTime.load` now raises `CoercionError` (instead of returning `nil`) if given values that can not coerced properly.
+* `Hash.dump` now first calls `Hash.load`, and correctly uses defined value types for dumping.
+* Added `Hash#get`, for retrieving keys using the same logic the `case_insensitive_load` and `allow_extra` with defined `extra` key.
 
 2.2.1
 ------